Move default_metadata_type_docs out of postgres driver into abstract.py
