+++ datacube/index/postgres/_metadata_types.py
@@ -4,24 +4,15 @@
 # SPDX-License-Identifier: Apache-2.0
 import logging
 import warnings
-from pathlib import Path
 
 from cachetools.func import lru_cache
 
 from datacube.index.abstract import AbstractMetadataTypeResource
 from datacube.model import MetadataType
-from datacube.utils import jsonify_document, changes, _readable_offset, read_documents
+from datacube.utils import jsonify_document, changes, _readable_offset
 from datacube.utils.changes import check_doc_unchanged, get_doc_changes
 
 _LOG = logging.getLogger(__name__)
-
-_DEFAULT_METADATA_TYPES_PATH = Path(__file__).parent.parent.joinpath('default-metadata-types.yaml')
-
-
-def default_metadata_type_docs():
-    """A list of the bare dictionary format of default :class:`datacube.model.MetadataType`"""
-    return [doc for (path, doc) in read_documents(_DEFAULT_METADATA_TYPES_PATH)]
-
 
 class MetadataTypeResource(AbstractMetadataTypeResource):
     def __init__(self, db):
